Fold RowWriter into result.go instead of a separate stream.go
RowWriter has one caller in this package (StreamRows) and no reason to
live in its own file; merging keeps all Result-shaped output code in
one place.
